refactor(ui): move chooseNewValue into its own module
Unchanged walk math; a separate file lets the unit tests import it without
pulling EffectsPanel's vanilla-extract styles into a node test run.
